Moved "DoneButton" to the parent directory
The "DoneButton" will be shared between all the new importer flows
